tildiary/requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.7
django-cors-headers==3.14.0
django-stubs==1.15.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==1.9.1
flake8==6.0.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy==1.0.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.7.1.2
types-PyYAML==6.0.12.8
types-requests==2.28.11.15
types-urllib3==1.26.25.8
typing_extensions==4.5.0
urllib3==1.26.14
